Paper over non-string error messages.
request-promise-core does not export its network exception types in the
public interface, and even if it did, it's not guaranteed they are the
only kinds of exception that might be thrown.

Handle these kinds of exceptions by simply assuming any non-string error
is a network error, and logging the real exception to the developer
console (Help -> Toggle Developer Tools) in case this isn't true.